pyproject.toml
[project]
name = "trucking-cost-pipeline"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyspark==4.1.1",
    "delta-spark==4.2.0",
    "yfinance>=1.3.0",
    "requests>=2.33.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.3", "ruff>=0.15.12", "jupyterlab>=4.5.2"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*"]
